+++ packages/components/pages/f-account-info/test-utils/component-objects/f-account-info-selectors.js
@@ -1,8 +1,8 @@
 exports.COMPONENT = '[data-test-id="account-info"]';
-exports.CHANGE_EMAIL_ADDRESS_LINK = '[id="change-email-address-link"]';
-exports.SAVE_CHANGES = '[data-test-id="account-info-save-changes-button"]';
-exports.CHANGE_PASSWORD = '[data-test-id="account-info-change-password-button"]';
-exports.DELETE_ACCOUNT = '[id="delete-account"]';
+exports.CHANGE_EMAIL_ADDRESS_LINK = '[name="account-info-change-email-link"]';
+exports.SAVE_CHANGES_BUTTON = '[data-test-id="account-info-save-changes-button"]';
+exports.CHANGE_PASSWORD_BUTTON = '[data-test-id="account-info-change-password-button"]';
+exports.DELETE_ACCOUNT_LINK = '[name="account-info-delete-account-link"]';
 
 
 exports.FIELDS = {